refactor integration test to be inside the specs directory and not in the main level
