+++ spec/integration/logstash_config/file_input_to_file_output_spec.rb
