+++ spec/integration/plugin_manager/plugin_install_spec.rb
