+++ spec/integration/plugin_manager/plugin_list_spec.rb
