+++ spec/integration/plugin_manager/plugin_uninstall_spec.rb
