+++ spec/integration/plugin_manager/plugin_update_spec.rb
