+++ spec/unit/license_spec.rb
