+++ spec/unit/plugin_manager/gemfile_spec.rb
