+++ spec/unit/plugin_manager/install_spec.rb
