+++ spec/unit/plugin_manager/update_spec.rb
