+++ spec/unit/plugin_manager/util_spec.rb
